api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
minio==7.2.8
python-multipart==0.0.9
pydantic==2.9.2
scikit-learn==1.5.2
numpy==1.26.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
httpx==0.27.2
aiofiles==24.1.0
python-dotenv==1.0.1
